Repackage wireguard service
